+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
-## [2.0.5] dev-2020-11-07
+## [2.0.5] dev-2020-11-09
 ### Added
 - New {ifingroup id|idnumber}{/ifingroup} tags.
 - New {filtercodes} tag. Note: Only works for teachers and above.
@@ -14,7 +14,8 @@ All notable changes to this project will be documented in this file.
 - New {timezone} tag.
 - New {preferredlanguage} tag.
 - New {coursesummary} tag.
-- New {firstaccessdate} tag/
+- New {firstaccessdate} tag.
+- New {formsesskey} tag.
 ### Updated
 - {courseprogress} and {courseprogressbar} now show zero progress if progress is 0.
 - {alert} to allow for optional contextual class stying.